pyproject.toml
[project]
name = "autos-ai"
version = "0.1.0"
description = "Agente de IA para análisis de solicitudes de seguros de flotilla de autos grupales"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Proprietary" }
authors = [{ name = "Autos AI Team" }]

dependencies = [
    # Google ADK - Agent Development Kit
    "google-adk>=0.5.0",
    # Google Cloud & Vertex AI
    "google-cloud-aiplatform>=1.70.0",
    "google-cloud-firestore>=2.19.0",
    "google-cloud-storage>=2.18.0",
    "google-cloud-documentai>=2.29.0",
    "google-cloud-bigquery>=3.25.0",
    "google-cloud-bigquery-storage>=2.25.0",
    "google-auth>=2.30.0",
    "vertexai>=1.0.0",
    # Document processing
    "pandas>=2.2.3",
    "openpyxl>=3.1.5",
    "pdfplumber>=0.11.4",
    "pypdf>=4.3.1",
    "python-docx>=1.1.2",
    "xlrd>=2.0.1",
    # API & Web
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    # Data validation & settings
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.1",
    # Utilities
    "python-multipart>=0.0.12",
    "aiofiles>=24.1.0",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
    "rich>=13.9.0",
    "typer>=0.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "black>=24.10.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "ipython>=8.29.0",
]

[project.scripts]
autos-ai = "app.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]


[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests", "eval/test_cases"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=app --cov-report=term-missing --cov-report=html"
